requirements.txt
google-api-python-client==2.187.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
python-dotenv==1.2.1
google-genai
